aws-secrets-controller: basic logging for secret reading time
We have some evidence that reading secrets from AWS might stall
indefinitely. Add a basic log line that indicates when an AWS secret
read starts and ends to prove or disprove this hypothesis if it occurs
again in production.

We log at an `info!` level so that we have visibility into every secrets
read. These are relatively infrequent, so this shouldn't be a meaningful
increase in log volume. We don't want to log at a lower level because
the hang has only occured rarely, so we want the logs on by default for
when it happens next.

Fix #27933.