storage: simplify command handling
The storage command handling code was somewhat convoluted, passing a
number of values (`internal_cmd_tx`, `async_worker`, `worker_index` to
`StorageState` methods that are already fields of `StorageState` and
therefore available through `self`.

This commit removes these additional arguments, simplifying the code a
bit in the process. It also removes the `Rc<RefCell<>>` wrapper for the
async worker, since nothing requires it. Finally, it makes `run_client`
not clone the `internal_cmd_tx` anymore, avoiding the risk of borrow
errors in that method.