storage: don't panic when dropping during reconciliation
This commit fixes a bug where storage would (soft) panic after a
reconciliation that caused the dropping of stale sources. The panic was
caused by the fact that the handler for the `AllowCompaction` command
verifies that specified IDs exist by looking them up in
`reported_frontiers` but reconciliation removes the IDs of dropped
objects from `reported_frontiers`.

The fix is to have reconciliation not send `AllowCompaction` commands to
drop storage collections, but instead to remove them from the
`StorageState` directly. To this end, a `drop_collection` method is
extracted that can be called from both the `AllowCompaction` command
handler and from reconciliation.