storage: continually advance upper of migrated builtin tables to track txns upper
Work towards zero-downtime upgrades: #27406

This makes it so that shards of migrated builtin tables remain
readable at a recent timestamp. We have to treat them separately because
we cannot yet register them in txns.

Importantly, we are careful not to advance the upper further than the
upper of the txn shard because those tables will have to be registered
in txns once a new deployment restarts in read-write mode.