catalog: permit altering certain system objects
Introduce a new `runtime_alterable` field for system connections, which
allows the connection to be altered by its owner. Because the system
object can drift from the SQL stored in the binary over time, migrations
are not triggered if the SQL in the binary is changed; the user's
definition is left in place. However, forcing a removal of the object by
removing it in a later version of Materialize is supported.

The system is designed to generalize to all types of system objects
(e.g., items, tables), but for simplicity we limit it to connections to
start, as that's the immediate need. The only runtime-alterable object
at the moment is the `mz_analytics` connection.