compute: track replica frontiers in `ReplicaCollectionState`
In contrast to all other per-replica, per-collection state, the compute
controller didn't track replica frontiers in
`self.replicas[].collections[].*_frontiers` but in
`self.collections[].replica_*_frontiers`. This has always been slightly
annoying but becomes very annoying when we are trying to introduce
per-replica `ReadHold`s for storage dependencies of compute collections.

This commit therefore refactors the compute controller to keep the
replica frontiers in the `ReplicaCollectionState`. This simplifies some
of the code, but note that it also temporarily removes the installation
of per-replica read holds. This is added in again in the next commit.

This commit also replaces some `tracing::error!` invocations with soft
panics, to make it more likely that we will notice bugs in CI.