adapter: JSON-format prohibited statement errors in the HTTP API
Report prohibited statement errors as we do query errors to improve the
consistency of this endpoint.

By storing the prohibited check result instead of exiting early we can
then eagerly execute the parsed queries in order until the first
unparsable query is reached, at which point it appends the parsing error
to the response object and returns (as we do with query errors).