environmentd: enable client-side mode in Segment client
Segment only records IPs for events on the server if the events
originate from a "client-side library" (e.g., a library that runs in the
user's browser). Segment determines the origin of an event based on the
`context.library.name` field. So, add an optional flag to environmentd
that causes our Segment client to emulate the Analytics.js library.
This flag is on by default in the evalution Docker image
(materialize/materialized).