Preserve the rest of the nullability info
Currently, the type of the field we generate depends on the specific
contents of the array. This means that two arrays generated by the same
encoder may not have the same type, which limits our ability to eg.
concatenate / consolidate them together later.

Instead, track and use nullability info at the encoder level