adapter: emit an object when session default cluster is missing
Previously, when a connection started with a default cluster that did
not exist, we'd only emit a warning if the setting came from the role or
the system. The removed comment within the patch explains the rationale.

Still, we can and should encourage the user to reconfigure their session
to use an extant cluster with `SET cluster`.