adapter: allow suppressing warnings for missing "current" objects
Materialize will generally attempt to warn the user if their session's
"current" database, schema, or cluster does not exist via the pgwire
notices.

However, not all applications want these notices. In particular, they're
quite spammy with dbt, which often establishes dozens of connections
with a known invalid cluster, which it ignores because it always runs
DDL commands with the cluster explicitly specified.

Fix #28857.