storage: render migrated builtin shards in txn-wal system
When handling shards minted for migrated collections, we want to render
their persist_sources in txn-wal mode, even though they're not yet
registered with the transaction shard. This ensures that, when the new
envd is promoted and registers the data shard with the transaction
shard, running persist_sources seamlessly transition to following the
new logical upper advancements. (Previously, they'd get stuck, because
they'd only be following the data shard.)

Fix #28919.