sql: add support for YUGABYTE connections and sources
Beyond the syntax level changes the underlying machinery remains
identical. The connection gained a `flavor` field that is used to track
whether we should expect a PostgreSQL or Yugabyte server on the other
side and is meant to be used by the actual operator to modify its
behavior accordingly.

The new syntax is guarded by a feature flag that we can enable for
everyone once all the pieces are in place.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>