adapter,balancer: log pgwire connection start/end
This commit assigns all pgwire connections a unique UUID and logs when
that connection starts and ends in the balancer and adapter.
Importantly, the UUID is propagated from the balancer to the adapter, so
we can correlate the connections across processes.  This will help us
debug broken connections.

Work towards resolving MaterializeInc/incidents-and-escalations#92.

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>