adapter: Log pgwire connection start (#28609)
This commit assigns all pgwire connections a unique UUID and logs when
that connection starts and stops in the balancer and adapter. This
will help us debug broken connections. The next step is to log when
the connection is severed, where it was severed and why.

Works towards resolving incidents-and-escalations/#92

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>