compute: replace dependents of replaced dataflows
This commit fixes a bug in compute reconciliation, which would reuse
dataflows even when dataflows they were reading from had been replaced.
This behavior was undesirable because the reused dataflows would keep
discarded dataflows alive, and the installed dataflows wouldn't match
the ones requested by the controller anymore.

To fix this we tweak the reconciliation logic to also replace dataflows
that depend on replaced dataflows.