compute: add as_of_selection module
This commit adds `mz_compute_client::as_of_selection`, a new module to
contain code responsible for bootstrap as-of selection.

This commit also adds some initial as-of selection logic, namely the
application of hard (i.e. correctness critical) constraints. The current
state is still not actually usable because it will select the empty
frontier for most dataflows.