persist: build percentage based dyncfg for structured part rollout
Further validate the structured data in customer environments by writing
some random small percentage of parts with both_v2. Doing this for all
parts is probably too expensive, particularly on read where we'd ~double
data volumes.

Then we'll turn on the row_with_validate to ensure that everything
round-trips losslessly though the new format. We should be able to do
this for 100% of parts that are written with both_v2, even in prod,
because there will be a small number of them.

Touches #24830