0dt: exit `environmentd` with exit code 0 when fenced
Resolves #29013

Before, we would exit with an error code that indicates a problem, and
we were using the same error code for cases where we want to be
restarted (restarting current deployment generation in read-write mode)
and cases where we don't need to be restarted (being fenced).

This will allow us to make alerting more fine grained and not restart
processes that are done, for example in tests.