tests: use on-failure in 0dt platform tests
We were using unless-stopped before, because of uncertainty around
on-failure working the way we expected. But I think we can switch now.

This also allows changing the "materialize native" restart policy in the
`materialized` image entrypoint to only be set on `on-failure`, which is
the correct thing to do after we recently changed that native restart
implementation to stop on exit code 0.