catalog: Remove storage usage from catalog
`mz_storage_usage_by_shard` is currently a builtin table that, like
most builtin tables, is backed by data in the durable catalog.
Specifically, it is backed by the storage usage collection. The
mz_storage_usage_by_shard table has the following columns:

  - `id uint8`
  - `shard_id text`
  - `size_bytes uint8`
  - `collection_timestamp timestamptz`

In order to have correct billing logic, we need the following guarantees:

  - `collection_timestamp` should be close to the real time that the
    row was written to the table.
  - collection_timestamp should be less than or equal to the logical
    timestamp that the row was written. i.e. If I run `SELECT * FROM
    mz_storage_usage_by_shard AS OF t` then I should see all rows that
    will ever exist s.t. `collection_timestamp <= t`.
  - collection_timestamp should be monotonically increasing.

The way we currently maintain these guarantees is by first generating
the `collection_timestamp` via the timestamp oracle, and then
generating the logical write timestamp via the timestamp oracle.

The storage usage collection takes up 99% of the space in the durable
catalog in all Materialize deployments. It takes a significant amount
of time to read this data into memory and iterate through it once it's
in memory. This commit attempts to resolve this issue by removing the
storage usage collection from the durable catalog. The source of truth
for this data will move from the durable catalog to the
`mz_storage_usage_by_shard`. This requires the following special
handling of the `mz_storage_usage_by_shard` table:

  - It is not cleared on startup.
  - We do not allow any migrations to the table, because that would
    clear the contents.

The startup process is still blocked on reading in the entire
`mz_storage_usage_by_shard` table and pruning old entries. A future
commit will move this pruning into a background task.

Works towards resolving #28722