adapter: remove wiring for manual dropping of read holds
This commit removes the adapter infrastructure
(`dropped_read_holds` channel, `DropReadHolds` message,
`release_read_holds` method) that was in place to ensure dropping of
compute read holds.

The plan is to take compute read holds as `ReadHold` objects, so we
won't need to drop them manually anymore. The commit does not yet
implement the move to `ReadHold`s for compute though, so it leaves
things in an unsound state for later commits to fix.