storage/sinks: remove `SinkRender::uses_keys`
We don't currently have sink implementations that return `false` for
`uses_keys`. So this commit remove the method.

Allowing this value to be configurable per-sink implementation added a
fair bit of complexity in a delicate part of the sink rendering code.
Since the code paths can't be exercised without a sink implementation
that doesn't use keys, there are likely lurking bugs, too.

If in the future we need to add support for a sink that doesn't support
keys, we can bring this logic back then, with tests.