storage/sinks/kafka: attach headers to tombstones
Due to a bug in the initial implementation of the `HEADERS` option for
Kafka sinks, we were not properly attaching headers to tombstones (i.e.,
messages with a `NULL` value).

The fix is to defer the final envelope encoding step until later in the
sink pipeline, so that the code that determines what headers to apply
can see the `before` value for tombstones.

This is a prerequisite to supporting custom partitioning (#28837).