persist: migrate structured arrow data as necessary on read
We now have both a write schema and a read schema at read time. The
write schema is specified by an id which indexes into a map in durable
persist state, but is sometimes missing because of backward
compatibility. The read schema is always present but might be missing an
id (e.g. if we use this mechanism for projection pushdown, but also for
other various edge cases).

The vast majority common case is that these schemas will be the same and
no migration is necessary. If we're missing the write schema, we _have_
to use the legacy Codec to decode the data. If we have both and they're
not the same, construct (and cache) migration logic for converting from
the write schema to the read schema.

Note that, counter-intuitively, the write schema could be from "after"
the read schema. This is because of the following scenario:

- A persist source is started on some shard at schema 0.
- The shard is migrated to schema 1. We don't want to restart the
  dataflow because it would be too disruptive.
- Data is written at schema 1 and the persist source encounters it.
- We solve this by only allowing migrations that are both backward and
  forward compatible with all registered migrations.
- In practice, this just means a field id cannot be reused after it is
  dropped. Nothing special to implement here right now because we don't
  (yet) support dropping columns in `compare_and_evolve_schema`.