cluster: use async channel to return activators
Previously, the channel through which cluster workers would return
dataflow activators to the tokio tasks managing client connections was a
sync channel. When the connection tasks called `recv` on them, that call
would synchronously block, stalling the current thread and even the
entire tokio runtime if it was single-threaded.

The fix is to use an async channel instead.