compute: wallclock lag summary metric
This commit improves on the existing dataflow wallclock lag metric by
making three changes:

 * Drops the `avg` aggregation in favor of total sum and total count.
 * Adjusts the metric names to immitate the summary metric type. We
   cannot use an actual summary metric because the `prometheus` crate
   doesn't support those.
 * Changes the aggregation approach to exposing the min/max value over
   the sliding window of the most recent 60 samples, instead of updating
   the metric values once per minute.

For the last part, a `MinMax` datastructure is introduced to efficiently
keep track of the min/max values are the sample window slides. Some care
is taken to reduce the memory usage of that datastructure to roughly 750
bytes per compute collection.