compute: assign empty as-ofs to sealed MVs
MV dataflows that sink into storage collections that have already
advanced to the empty frontier don't need to produce any output anymore,
so we can optimize by assigning them empty as-ofs.

The implementation does so by removing them from the set of collections
tracked during the as-of selection process, which has the benefit of
also making this process more efficient, since it has to consider fewer
collections now.