sql: remove require_feature_flag_w_dynamic_desc
The only use case for this was limiting access to WITH options for
CREATE SOURCE and CREATE SINK, but we no longer have any WITH options
that are feature flagged.

In the future, if we need to limit use of a WITH option, we can do so
with a per-option feature flag, rather than using the denylist
approach that was removed here.