sql: improve error message for private preview features
Today if a user tries to use a feature in private preview, they see the
following error message:

    ERROR: <FEATURE> is unsupported
    DETAIL: The requested feature is typically meant only for internal development and testing of Materialize.

This is plain wrong for most feature flags, which gate access to a
feature that's meant to be user facing eventually and is just in private
preview for the moment. We want to actively encourage users to reach out
about enabling these features in their account.

This commit adjusts the error message accordingly:

    ERROR:  LIMIT <expr> syntax is not available
    DETAIL:  The requested feature (enable_expressions_in_limit_syntax) is in private preview.
    HINT:  Contact support to discuss enabling the feature in your Materialize region.

One callout: we always include the name of the feature flag in the error
message. The previous error message would only show the feature flag
name if running locally with `--unsafe-mode` enabled. But the feature
flag names aren't a secret. Including them in the error message
unconditionally will make it easier to find the feature flag that needs
to be enabled when a customer copies/pastes the error message into a
support ticket.