adapter,sql: mark unsafe feature flags as such
Several feature flags gate access to permanent or near-permament escape
hatches that we never plan to enable in production. Mark those feature
flags as unsafe. This prevents us from accidentally enabling the
features in production, and also ensures the error message properly
indicates the situation...

    materialize=> create table t (a int primary key);
    ERROR:  CREATE TABLE with a primary key or unique constraint is not supported
    DETAIL:  The requested feature (unsafe_enable_table_keys) is unsafe and is meant only for internal development and testing of Materialize.

...rather than claiming that the feature is in private preview and that
it may possible to join the preview by talking to support.