adapter,pgwire: use conn UUID as session UUID; display in welcome message
Use the connection UUID added in 3b42b8ec3, which tracks connections
across balancerd and environmentd processes, as the session UUID. Then
show the session UUID in the welcome message.

This will make it easier to correlate particular connections with their
environmentd and balancerd logs, which will help in tracking down
MaterializeInc/incidents-and-escalations#92.