cluster: avoid accidentally dropping TimelyContainer
This commit fixes a cancel safety bug: Previously, the `ClusterClient`
would `take` the `TimelyContainer` out of the `Arc`-shared state and
then await a response from a channel. If the GRPC client disconnected
while that await was still pending, the `ClusterClient` would be dropped
and with it the `TimelyContainer`. Dropping a `TimelyContainer` makes is
synchronously block on the Timely workers to finish, which never
happens, so the entire tokio thread becomes blocked forever.

The fix implemented here if to never move the `TimelyContainer` out of
its `Arc`. Additionally, we add a panicking `Drop` impl for
`TimelyContainer`, to ensure we can quickly diagnose similar bugs in the
future.

This is only a quick-fix. Notably, it doesn't ensure that the stream
created by `forward_bidi_stream` is cancel safe, it only removes one
specific instance of cancel safety. There might or might not be other
cancel safety bugs left. A proper fix should include a refactor that
brings the code into a state where cancel safety is easy to verify.