persist: add backward compatible schema detection and migration (#29046)
This works on the arrow::DataType of the output produced by the two
`Schema2` encoders. It detects when some `new` schema is backward
compatible with an `old` one (add+remove nullable column, including in
nested structs). If so, it returns logic that migrates arrays produced
by `old` to have the same DataType as arrays produced by `new`.

The returned logic knows when the migration involves dropping data as
well as whether the migration preserves array order (though only the
former is initially implemented in this PR).

This will be used in three places:
- When registering a new schema for a shard, we'll require that it be
backward compatible with all previous schemas. Being able to drop
columns complicates this, and is not necessary for the initial scope of
the mz schema change work, and so we disallow it for now.
- When reading data, we'll allow the reader to request projection
pushdown by providing a schema with only the columns it is interested
in.
- Compaction will want to operate on inputs with matching schemas.
Exactly what this looks like is still TBD, but it will certainly want to
know if migrations it uses to do this are order preserving (e.g.
dropping column B from `A, B, C` is not).

Touches #16625