environmentd: gracefully terminate when fenced out by new gen
If an environmentd process discovers that it's been fenced out by an
environmentd process with a newer generation, it should exit with code
zero (0). The entrypoint.sh script in the environmentd Docker image will
then sleep until the orchestration layer gets around to tearing down the
container.

Previously in this situation, the old environmentd would halt with exit
code two (2), then restart and panic upon discovering that the catalog
had been written to by a newer version. It would then remain in a panic
loop until the orchestration layer got around to tearing down the
container. This did not cause incorrectness, but led to log spam in
tests and noisy alerts during release windows in staging and production.

Fix #29199.