adapter: Avoid necessary catalog clones (#29210)
The `Coordinator` keeps the `Catalog` behind an `Arc` reference. It
occasionally hands out references to the `Catalog` by cloning the `Arc`.
When it wants to mutate the `Catalog` it calls
`Coordinator::catalog_mut()`, which under the hood calls
`Arc::make_mut()`. This function will clone the catalog iff any other
`Arc` references to the catalog exists.

During startup the `Coordinator` was accidentally creating a reference
to the catalog and then immediately calling `catalog_mut()` for every
index and materialized view. As a result, the catalog was cloned once
per index and materialized view.

This commit drops the reference to the catalog before calling
`catalog_mut()` preventing the catalog from being cloned. Locally this
causes each call to `catalog_mut()` to speed up from 20-30 ms to 5-10
ns.

Works towards resolving #28722