bin/cloud-push: use proper prerelease versions
Rather than using a random tag, use a tag that includes the current
version from Cargo.toml. That way, the cloud orchestration layer will be
able to understand precisely what features (e.g., CLI arguments) this
custom build supports.

(Of course, if the custom build adds *new* features that are meant to be
understood by the cloud orchestration layer, that will require bumping
the version in Cargo.toml; that's fundamental to this approach of using
prerelease versions to gate features that impact the cloud orchestration
layer, though.)