catalog: Improve fencing
Previously, the durable catalog could be fenced by one of the
following ways:

  - It tried to listen to the durable catalog and saw a higher epoch.
  - It tried to listen to the durable catalog and saw a higher binary
    version.
  - It tried to write to the durable catalog and had an upper
    mismatch.
  - It tried to write to the catalog migration shard and had an upper
    mismatch.

These would either result in a non-descriptive
`DurableCatalogError::Fence` error or a
`DurableCatalogError::IncompatiblePersistVersion` error. Additionally,
if multiple of these conditions happened at the same time, it was
non-deterministic which error would be returned.

This commit modifies the conditions for when the durable catalog can be
fenced to the following conditions:

  - It tried to listen to the durable catalog and saw a higher deploy
    generation.
  - It tried to listen to the durable catalog and saw a higher epoch.
  - It tried to write to the durable catalog shard and had an upper
    mismatch.
  - It tried to write to the catalog migration shard and had an upper
    mismatch.

When initializing the catalog, we still check that the current binary
version is not less than the catalog binary version, but we no longer
check during every update. After the catalog has been initialized,
another catalog can only increase the binary version by also fencing
the current catalog by another method.

This commit also creates a new error enum to explicitly codify these
different types of fences. When multiple of these conditions happen at
the same time, then we deterministically will return a fence error with
the priority of how the conditions are ordered above. The context of
the fence is useful to higher layers in determining what action to
take. For example if the deploy generation has increased, then we'd
like to gracefully exit with exit code 0 and let the new instance take
over.

Works towards resolving #29199