compute: don't expose collections' implied capability
This commit removes `CollectionState::read_capability`, a public method
that exposed the implied capability (from `implied_read_hold`) to
clients of compute. There has in the past been confusion about the
semantics of this method, especially compared to `read_frontier`. Since
there is currently no reason for clients to care about the implied
capability, we can remove the method to resolve that confusion.