adapter,controller: mz_cluster_replica_metrics source
This commit converts `mz_cluster_replica_metrics` from a "retained
metrics" builtin table to an append-only builtin source. The coordinator
no longer needs to be involved in writing metrics data to storage, this
can now happen in the controller.

This commit does not include any code to partially truncate
`mz_cluster_replica_metrics`, which is something we want to ensure it
doesn't grow unboundedly. A subsequent commit will deal with that.