adapter: correctly calculate the minimum wait time
The variable changed in this PR is supposed to track the shortest wait
needed for the time to advance to the time of the timestamp oracle. It
is updated as part of processing the list of requests by continually
adjusting it via `shortest_wait = min(shortest_wait, wait)`.
Unfortunately since the value was initialized at the minimum value
(instead of the MAX) this code would always calculate a duration of
zero, and therefore if it somehow triggered the database would spin as
fast as possible generating requests for linearization which in some
tests would take minutes to clear, all while being totally unresponsive.

The fix is to initialize the variable to the maximum duration.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>