catalog: Retry fences
There is a race when one catalog process is trying to fence another catalog
process. The race looks like this:

  1. Catalog B prepares to write a new epoch to the durable catalog.
  2. Catalog A commits a write to the durable catalog (i.e. some DDL).
  3. Catalog B fails to commit the prepared write with an upper
     mismatch error.

Previously, the catalog would return an upper mismatch error back to
the caller and force them to deal with it. Usually that involved
re-opening the catalog from scratch, which is an expensive operation.

This commit fixes the issue by retrying the fence internally within the
durable catalog. This is safe to do because before fencing, the durable
catalog has not done anything significant. Upgrades/migrations all
happen after the fence; fencing is the first thing the catalog does
after reading all updates into memory.