Peeks against sealed collections in serializable mode can't step forward
Remove a warning that will happen when querying a collection that has its
write frontier advanced to [] and querying it in serializable isolation
level. This means we'll select an as-of of the maximal timestamp, which we
cannot step forward, hence the warning fires. Setting the until to the
empty frontier (which it is by default) is the correct action in this case.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>