catalog,controller: add replica metrics history
This commit adds a new builtin source,
`mz_cluster_replica_metrics_history`, and a corresponding view
`mz_cluster_replica_utilization_history` on top of it. They contain the
data already present in `mz_cluster_replica_{metrics,utilization}`, but
as (pseudo-) append-only collections, rather than retained-metrics
collections.

This plan is to remove the old retained-metrics collections in favor of
the new append-only ones, but we need to wait until the new ones had
time to backfill the historical data present in the old ones.

This commit does not include any code to partially truncate
`mz_cluster_replica_metrics_history`, which is something we want to
ensure it doesn't grow unboundedly. A subsequent commit will deal with
that.