storage: truncate replica metrics history
This commit adds truncation of the replica metrics history collection
during bootstrap. In contrast to the existing status history
collections, this collection is truncated based on the age of events
instead of the number of events per ID.

The implementation of the truncation method is kept simple. In the
future it can be refined to avoid reading the entire collection into
memory, if it turns out that the size of a metrics collection is
problematic.