persist: fix bug where non-struct stats caused us to skip columnar data
`encode_updates` returned an `Err` if either it was unable to encode the
data into structured arrow data OR if it was unable to turn the stats
into a non-nullable `StructStats`. It turns out that all persist unit
tests were returning an unexpected stats type, and so we essentially had
much less test coverage than we thought. We were logging an `error!`
(thankfully, it doesn't seem to be showing up in sentry), but this
wasn't loud enough that we noticed until now.

So, replace the error log with a `soft_assert_or_log!` so that tests
(and CI) fail loudly when something unexpected happens. Then fix the new
failures :). We could make non-struct stats roundtrip, but it'd be Work
and Code and unused in production, so instead require `ColumnDecoder`
implementations to figure out how to map themselves into `StructStats`.
We can always extend this support later, if needed.