adapter,controller: exclude new builtins from hydration check
This commit adjusts the 0dt hydration check to exclude newly created
builtin objects. These objects are sometimes not able to hydrate because
neither the old nor the new environment is writing to them. Excluding
them from the check is fine: Clients connected to the old environment
don't know of their existence, so no downtime is induced when they are
not immediately hydrated after the cutover.