catalog: Clean up some tests
Previously, a lot of catalog tests had copy and pasted `NOW_ZERO()` as
the clock and a comment explaining why we were using that clock. The
comment is not actually valid for the majority of the tests, so this
commit removes the comment and switches to `SYSTEM_CLOCK()` which is
what the catalog uses in production.