adapter: move storage usage pruning to bootstrap
Storage usage pruning blocks startup, but the time spent pruning was
hidden from the `envd_start_duration` metric by being in `coord.serve`
rather than `coord.bootstrap`.

We would prefer to move this to a background task that's kicked off by
`serve`, but there are complications with running this in the
background, and moving it from `bootstrap` to `serve` later is not the
hard part.