catalog: Fix invalid fence updates
When fencing other processes, the catalog will retry writing the new
epoch until it is successful. Previously, the catalog would
accidentally update it's internal epoch related state on every retry.
This would cause the catalog to incorrectly think that the existing
epoch was larger than it actually was. This commit fixes the issue by
only updating the epoch related state after the epoch write was
successful.

Fixes #29293