cluster: singleton `ClusterClient`
Previously, each client connecting to a cluster via GRPC would build a
new `ClusterClient` and connect it to the existing Timely cluster. This
introduced a bunch of complexity in the code, as well as some tricky
interactions, for example when multiple clients were connected to the
same Timely cluster, or when a client connection was blocked waiting to
receive an activator from an unresponsive Timely cluster.

Now there is only a single `ClusterClient` that is held by one GRPC
client at a time and owns all access to the Timely cluster. New GRPC
clients need to wait for existing clients to observe their cancellation
and exit before they can obtain the `ClusterClient`. On the other hand,
they don't need to wait for the Timely cluster to send an activator
anymore, which should make GRPC clients immune to unresponsiveness in
the Timely cluster.

There is still one defect, namely that new clients can observe compute
responses meant for previous clients, which would confuse the controller
if they would be forwarded. A subsequent commit will fix this.