catalog: Add new fence token variant
This commit adds a new variant to the catalog called `FenceToken`.
The fence token is a combination of the epoch and deploy generation so
that they can be written and read atomically in a single write.
Currently, this variant is not used.

The epoch is needed for bootstrapping and is therefore needed before
running catalog migrations. Therefore, this migration must span two
releases. In the first release we introduce the new variant. Then in
the second release we replace all epochs and deploy generations with a
fence token before the migrations have actually run.