storage/sinks/kafka: don't wedge on empty progress topics
Previously, Kafka sinks would wedge on startup if their progress topic
was empty but had a nonzero low water mark. It's extremely difficult for
a progress topic to get into this state (you need to write one or more
messages to the progress topic and then somehow delete *all* of those
messages), but we've observed at least one progress topic in this state
in the wild.

So, this commit teaches Kafka sinks to handle this case without wedging,
and adds a test case to prevent future regressions. See the comments
within the patch and the test case for details.