kafka-util: only complain when true misconfiguration happens
The code as it was would produce an error report if someone set the
kafka transaction timeout to a value that is above the default progress
record fetch timeout, even though it's the system that chose the
problematic default in the first place. So in some sense, it was
complaining about a problem it caused to itself.

This PR makes it so that configurations whose defaults are derived from
other values have an `Option<_>` type so that we can differenciate
between "the user does not have an opinion" and the "the user has an
opinion". We then only produce an error if the user had the wrong
opinion.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>